Move IR-related files to `librustc_codegen_ironox/ir`.
