+++ src/librustc_codegen_ironox/ir/basic_block.rs
@@ -7,15 +7,7 @@
 // except according to those terms.
 
 use context::CodegenCx;
-use value::Value;
-
-macro_rules! asm {
-    ($m:expr, $($args:expr)*) => {
-        $(
-            $m.instrs.push(format!("{}\n", $args));
-        )*
-    }
-}
+use value::{Instruction, Value};
 
 /// The index of the parent function, and the index of the basic block
 /// in the function.
@@ -27,8 +19,8 @@ pub struct BasicBlock(pub usize, pub usize);
 pub struct BasicBlockData {
     /// The label of the basic block.
     pub label: String,
-    /// The x86-64 instructions in this basic block.
-    pub instrs: Vec<String>,
+    /// The instructions in this basic block.
+    pub instrs: Vec<Instruction>,
     /// The function the basic block belongs to.
     pub parent: Value,
     /// The terminator of the basic block.
@@ -43,8 +35,6 @@ impl BasicBlockData {
             parent: parent,
             terminator: None,
         };
-        asm!(bb,
-             format!("{}:", label));
         let parent = match parent {
             Value::Function(p) => p,
             _ => bug!("The parent of a basic block has to be a function")
