+++ src/librustc_codegen_ironox/ir/struct_.rs
@@ -1,0 +1,16 @@
+use value::Value;
+
+/// A `struct` type.
+#[derive(Debug)]
+pub struct IronOxStruct {
+    /// The fields of the `struct`.
+    components: Vec<Value>,
+}
+
+impl IronOxStruct {
+    pub fn new(components: &[Value]) -> IronOxStruct {
+        IronOxStruct {
+            components: components.to_vec()
+        }
+    }
+}
